+++ supreme/src/commonMain/kotlin/at/asitplus/signum/supreme/symmetric/KeyGen.kt
@@ -3,18 +3,22 @@ package at.asitplus.signum.supreme.symmetric
 import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.signum.HazardousMaterials
+import at.asitplus.signum.indispensable.asn1.encoding.bitLength
 import at.asitplus.signum.indispensable.symmetric.CipherKind
 import at.asitplus.signum.indispensable.symmetric.Nonce
 import at.asitplus.signum.indispensable.symmetric.SymmetricEncryptionAlgorithm
 import at.asitplus.signum.indispensable.symmetric.SymmetricKey
-import at.asitplus.signum.indispensable.asn1.encoding.bitLength
+import org.kotlincrypto.SecureRandom
 import kotlin.jvm.JvmName
 
+private val secureRandom = SecureRandom()
+
+
 fun SymmetricEncryptionAlgorithm<*, *>.randomKey(): SymmetricKey<*, *> =
-    encryptionKeyFrom(secureRandom.nextBytesOf(keySize.bytes.toInt())).getOrThrow()
+    keyFrom(secureRandom.nextBytesOf(keySize.bytes.toInt())).getOrThrow()
 
 @JvmName("randomKeyWithNonce")
-fun <A: CipherKind>SymmetricEncryptionAlgorithm<A, Nonce.Required>.randomKey(): SymmetricKey<A, Nonce.Required> =
+fun <A : CipherKind> SymmetricEncryptionAlgorithm<A, Nonce.Required>.randomKey(): SymmetricKey<A, Nonce.Required> =
     (this as SymmetricEncryptionAlgorithm<*, *>).randomKey() as SymmetricKey<A, Nonce.Required>
 
 @JvmName("randomKeyUnauthenticated")
@@ -59,64 +63,64 @@ internal fun SymmetricEncryptionAlgorithm<*, Nonce.Required>.randomNonce() =
     @OptIn(HazardousMaterials::class) secureRandom.nextBytesOf((nonce.length.bytes).toInt())
 
 
-fun SymmetricEncryptionAlgorithm<*, *>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<*, *>> =
+fun SymmetricEncryptionAlgorithm<*, *>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<*, *>> =
     catching {
-        require(keyBytes.size == this.keySize.bytes.toInt()) { "Invalid key size: ${keyBytes.size * 8}. Required: keySize=${keyBytes.size.bitLength}" }
+        require(bytes.size == this.keySize.bytes.toInt()) { "Invalid key size: ${bytes.size * 8}. Required: keySize=${bytes.size.bitLength}" }
         @OptIn(HazardousMaterials::class)
         when (this.cipher) {
             is CipherKind.Authenticated.Integrated, is CipherKind.Unauthenticated -> SymmetricKey.Integrated(
                 this,
-                keyBytes
+                bytes
             )
 
             is CipherKind.Authenticated.WithDedicatedMac<*, *> -> SymmetricKey.WithDedicatedMac(
                 this as SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>,
-                keyBytes
+                bytes
             )
         }
     }
 
 
 @JvmName("fixedKeyWithNonce")
-fun SymmetricEncryptionAlgorithm<*, Nonce.Required>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<*, Nonce.Required>> =
-    (this as SymmetricEncryptionAlgorithm<*, *>).encryptionKeyFrom(keyBytes) as KmmResult<SymmetricKey<*, Nonce.Required>>
+fun SymmetricEncryptionAlgorithm<*, Nonce.Required>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<*, Nonce.Required>> =
+    (this as SymmetricEncryptionAlgorithm<*, *>).keyFrom(bytes) as KmmResult<SymmetricKey<*, Nonce.Required>>
 
 @JvmName("fixedKeyUnauthenticated")
-fun SymmetricEncryptionAlgorithm<CipherKind.Unauthenticated, *>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Unauthenticated, *>> =
-    (this as SymmetricEncryptionAlgorithm<*, *>).encryptionKeyFrom(keyBytes) as KmmResult<SymmetricKey<CipherKind.Unauthenticated, *>>
+fun SymmetricEncryptionAlgorithm<CipherKind.Unauthenticated, *>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Unauthenticated, *>> =
+    (this as SymmetricEncryptionAlgorithm<*, *>).keyFrom(bytes) as KmmResult<SymmetricKey<CipherKind.Unauthenticated, *>>
 
 @JvmName("fixedKeyUnauthenticatedWithNonce")
-fun SymmetricEncryptionAlgorithm<CipherKind.Unauthenticated, Nonce.Required>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Unauthenticated, Nonce.Required>> =
-    (this as SymmetricEncryptionAlgorithm<*, *>).encryptionKeyFrom(keyBytes) as KmmResult<SymmetricKey<CipherKind.Unauthenticated, Nonce.Required>>
+fun SymmetricEncryptionAlgorithm<CipherKind.Unauthenticated, Nonce.Required>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Unauthenticated, Nonce.Required>> =
+    (this as SymmetricEncryptionAlgorithm<*, *>).keyFrom(bytes) as KmmResult<SymmetricKey<CipherKind.Unauthenticated, Nonce.Required>>
 
 @JvmName("fixedKeyAuthenticatedIntegrated")
-fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.Integrated, *>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, *>> =
-    (this as SymmetricEncryptionAlgorithm<*, *>).encryptionKeyFrom(keyBytes) as KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, *>>
+fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.Integrated, *>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, *>> =
+    (this as SymmetricEncryptionAlgorithm<*, *>).keyFrom(bytes) as KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, *>>
 
 @JvmName("fixedKeyAuthenticatedIntegratedWithNonce")
-fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.Integrated, Nonce.Required>.encryptionKeyFrom(keyBytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, Nonce.Required>> =
-    (this as SymmetricEncryptionAlgorithm<*, *>).encryptionKeyFrom(keyBytes) as KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, Nonce.Required>>
+fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.Integrated, Nonce.Required>.keyFrom(bytes: ByteArray): KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, Nonce.Required>> =
+    (this as SymmetricEncryptionAlgorithm<*, *>).keyFrom(bytes) as KmmResult<SymmetricKey<CipherKind.Authenticated.Integrated, Nonce.Required>>
 
 @JvmName("fixedKeyAuthenticatedWithDedicatedMacWithNonce")
-fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, Nonce.Required>, Nonce.Required>.encryptionKeyFrom(
-    keyBytes: ByteArray,
+fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, Nonce.Required>, Nonce.Required>.keyFrom(
+    bytes: ByteArray,
     dedicatedMacKeyOverride: ByteArray? = null
 ): KmmResult<SymmetricKey.WithDedicatedMac<Nonce.Required>> =
-    (this as SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>).encryptionKeyFrom(
-        keyBytes,
+    (this as SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>).keyFrom(
+        bytes,
         dedicatedMacKeyOverride = dedicatedMacKeyOverride
     ) as KmmResult<SymmetricKey.WithDedicatedMac<Nonce.Required>>
 
 @JvmName("fixedKeyAuthenticatedWithDedicatedMAC")
-fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>.encryptionKeyFrom(
-    keyBytes: ByteArray,
+fun SymmetricEncryptionAlgorithm<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>.keyFrom(
+    bytes: ByteArray,
     dedicatedMacKeyOverride: ByteArray? = null
 ): KmmResult<SymmetricKey<CipherKind.Authenticated.WithDedicatedMac<*, *>, *>> = catching {
-    require(keyBytes.size == this.keySize.bytes.toInt()) { "Invalid key size: ${keyBytes.size * 8}. Required: keySize=${keyBytes.size.bitLength}" }
+    require(bytes.size == this.keySize.bytes.toInt()) { "Invalid key size: ${bytes.size * 8}. Required: keySize=${bytes.size.bitLength}" }
     @OptIn(HazardousMaterials::class)
     SymmetricKey.WithDedicatedMac<Nonce>(
         this,
-        secretKey = keyBytes,
-        dedicatedMacKey = dedicatedMacKeyOverride ?: keyBytes
+        secretKey = bytes,
+        dedicatedMacKey = dedicatedMacKeyOverride ?: bytes
     )
 }
